hw garbage collection - array

+++ gradle/hw08-gc/src/main/java/ru/otus/CalcDemo.java
@@ -15,7 +15,8 @@ import java.time.LocalDateTime;
 
 public class CalcDemo {
     public static void main(String[] args) {
-        long counter = 100_000_000;
+//        long counter = 100_000_000;
+        int counter = 100_000_000;
         var summator = new ru.otus.Summator();
         long startTime = System.currentTimeMillis();
